use result of nameFilter search
block-adj.js : split thenOrNow() out to form promises.js
manage-explorer.js : wrap datasetsP with thenOrNow() because result of datasetOrBlockMatch() is not a promise.
frontend/app/ :
 59b9a4d 37146 Jan 27 20:58  components/draw/block-adj.js
 4994220 50655 Jan 27 20:58  components/panel/manage-explorer.js
 7068458   367 Jan 27 20:58  utils/common/promises.js	(added)
